revert changes to Amazon's  declarations

+++ src/com/amazon/creators/model/BrowseNode.php
@@ -49,7 +49,7 @@ class BrowseNode implements ModelInterface, ArrayAccess, \JsonSerializable
         'displayName' => 'string',
         'id' => 'string',
         'isRoot' => 'bool',
-        'salesRank' => 'int'
+        'salesRank' => 'float'
     ];
 
     /**
